Lock down access to teams, spaces and resources based on team memberships and roles
Authorization is carried out at the controller level only, with appropriate conditions in the UI to determine when to show actions/links/etc.

Specs are included to check the authorization of every endpoint.

Uses [CanCanCan](https://github.com/CanCanCommunity/cancancan) to perform the authorization.

Rules:
- Only hub admins can make/revoke hub admins
- Only hub and team admins can edit (their) teams
- Any logged in user can see the list of all teams and spaces in the hub
  - … but only ones they have access to are clickable in the UI
- Any logged in user can create a new team (and automatically becomes team admin)
- Any team member can view the team page for their teams
- Only hub and team admins can edit and delete teams
  - + lock down UI
- Only hub and team admins can add/remove users from teams
  - + lock down UI
- Only hub and team admins can make/revoke team admins
  - + lock down UI
- Only hub admins and any team member can view a space page within a team
- Only hub admins and any team member can create, edit and delete a space within a team
- Only hub admins and any team member can create and delete resources within spaces in a team
- Only hub admins and any team member can run the resources bootstrap within spaces in a team
- Change hub admin controllers to now use the new authorization mechanism
- Skip authorization for:
  - All `/me` endpoints (since they always acts on the current user anyway)
  - Teams index page
  - Spaces index page
  - Healthcheck

+++ app/helpers/teams_helper.rb
@@ -4,6 +4,8 @@ module TeamsHelper
   end
 
   def delete_team_link(team, css_class: nil)
+    return unless can?(:destroy, team)
+
     link_to 'Delete',
       team_path(team),
       method: :delete,
@@ -42,8 +44,12 @@ module TeamsHelper
       class: css_class,
       data: {
         params: { role: role }.to_param,
+        'disable-with': 'Processing...',
         turbolinks: false,
-        'disable-with': 'Processing...'
+        confirm: 'Are you sure you want to update the role for this team member?',
+        title: "Update role for team member: #{user.email}",
+        verify: 'yes',
+        verify_text: "Type 'yes' to confirm"
       },
       role: 'button'
   end
